+++ scripts/mcp-tool-pack-baseline.json
@@ -8,12 +8,10 @@
   "admin_run_seed",
   "admin_view_logs",
   "approve_decomposition",
-  "assess_archetype_refinement",
   "assess_complexity",
   "assess_contribution",
   "check_deployment_windows",
   "check_sandbox",
-  "confirm_taxonomy_placement",
   "contribute_to_hive",
   "create_backlog_item",
   "create_build_epic",
@@ -30,7 +28,6 @@
   "evaluate_tool",
   "execute_promotion",
   "generate_code",
-  "generate_custom_archetype",
   "get_backlog_item",
   "get_build_dispatch_history",
   "get_build_engine_readiness",
@@ -48,7 +45,6 @@
   "list_epics",
   "list_patch_posture",
   "list_sandbox_files",
-  "prefill_onboarding_wizard",
   "principle_decide",
   "process_backlog_for_build_studio",
   "promote_to_build_studio",
@@ -95,7 +91,6 @@
   "start_sandbox",
   "start_scout_research",
   "submit_feedback",
-  "suggest_taxonomy_placement",
   "summon_coworker",
   "triage_backlog_item",
   "trigger_contributor_inventory_sync",
